+++ src/Capability.php
@@ -41,11 +41,12 @@ class Capability implements Serializable, JsonSerializable
 	 * @param string $capability
 	 * @param array $default_operators To apply if operations were not set.
 	 * @since 1.0.0
+	 * @since 1.0.1 Default operations or empty array to $this->_operations
 	 * @return void
 	 */
 	public function __construct ( string $capability = null, array $default_operators = null )
 	{
-		$this->_operations = $default_operators ?? ['any'];
+		$this->_operations = $default_operators ?? [];
 
 		if ( empty($capability) )
 		{ return; }
@@ -96,8 +97,53 @@ class Capability implements Serializable, JsonSerializable
 			return $this; 
 		}
 
-		$this->_operations = \explode(',', $_blocks['operators']);
+		$_blocks['operators'] = \explode(',', $_blocks['operators']);
+
+		// Has all available operations, so adding any instead
+		if ( CapabilityOperations::hasAll($_blocks['operators']) )
+		{ 
+			$this->_operations = ['any'];
+			return $this; 
+		}
+
+		$this->_operations = $_blocks['operators'];
 		return $this;
+	}
+
+	/**
+	 * Import Capability from array. Following the schema:
+	 * [ $key => $operations ].
+	 *
+	 * @param array $caps
+	 * @since 1.0.0
+	 * @return void
+	 */
+	public static function fromArray ( array $caps ) : Capability
+	{
+		foreach ( $caps as $key => $operations )
+		{ return (new Capability())->setKey($key)->set($operations); }
+
+		return null;
+	}
+
+	/**
+	 * Import Capability from json.
+	 * { "$key": [$operations] }.
+	 *
+	 * @param string $json
+	 * @since 1.0.1
+	 * @return void
+	 * @throws InvalidArgumentException
+	 */
+	public static function fromJson ( string $json ) : Capability
+	{ 
+		$json = json_decode($json, true);
+
+		if ( JSON_ERROR_NONE !== \json_last_error() ) 
+		{ throw new InvalidArgumentException(\sprintf('Cannot parse Capability object from JSON: `%s`.', \json_last_error_msg())); }
+
+		$cap = (new Capability())->setKey(\array_key_first($json));
+		return $cap->set($json[$cap->getKey()]);
 	}
 
 	/**
@@ -150,17 +196,42 @@ class Capability implements Serializable, JsonSerializable
 
 	/**
 	 * Add a new operation.
-	 * If current Capability has the "any" operation,
-	 * which allows any... then it will removes.
+	 * 
+	 * If current Capability allow "any" operation,
+	 * will thrown an exception to use `disallowAny`
+	 * method first.
 	 *
 	 * @param string|array ...$operations
 	 * @since 1.0.0
+	 * @since 1.0.1 Thrown an exception if cannot add because capability allow any.
 	 * @return self
+	 * @throws InvalidArgumentException If cannot add because capability allow any.
 	 */
 	public function add ( ...$operations )
 	{ 
 		if ( $this->isAnyAllowed() )
-		{ $this->_flushAny(); }
+		{ throw new InvalidArgumentException('The capability allows `any` operations. You should use `disallowAny()` first to adding new operations.'); }
+
+		$this->merge($operations);
+		return $this; 
+	}
+
+	/**
+	 * Insert a new operation.
+	 * 
+	 * If current Capability allow "any" operation,
+	 * flushes operations by removing it an continue
+	 * adding.
+	 *
+	 * @param string|array ...$operations
+	 * @since 1.0.0
+	 * @since 1.0.1 Thrown an exception if cannot add because capability allow any.
+	 * @return self
+	 */
+	public function insert ( ...$operations )
+	{ 
+		if ( $this->isAnyAllowed() )
+		{ $this->_operations = []; }
 
 		$this->merge($operations);
 		return $this; 
@@ -171,12 +242,18 @@ class Capability implements Serializable, JsonSerializable
 	 *
 	 * @param array $operations
 	 * @since 1.0.0
+	 * @since 1.0.1 Add "any" if $this->_operations has all operations available.
 	 * @return self
 	 * @throws InvalidArgumentException
 	 */
 	public function merge ( array $operations )
 	{ 
 		$this->_operations = \array_unique( \array_merge( $this->_operations, $this->_fixSpread($operations) ) );
+		
+		// If has all available operations, change it to any
+		if ( CapabilityOperations::hasAll($this->_operations) )
+		{ $this->_operations = ['any']; }
+		
 		return $this; 
 	}
 
@@ -185,12 +262,23 @@ class Capability implements Serializable, JsonSerializable
 	 * 
 	 * @param string $operations
 	 * @since 1.0.0
+	 * @since 1.0.1 If any is allowed, add all available operations removing $operations.
 	 * @return self
 	 * @throws InvalidArgumentException
 	 */
 	public function remove ( ...$operations )
 	{
-		$this->_operations = \array_diff($this->_operations, $this->_fixSpread($operations));
+		$operations = $this->_fixSpread($operations);
+
+		// If any is allowed, then get all operations allowed
+		// and remove required $operations from it
+		if ( $this->isAnyAllowed() )
+		{ 
+			$this->_operations = \array_diff(CapabilityOperations::operations(), $operations); 
+			return $this;
+		}
+
+		$this->_operations = \array_diff($this->_operations, $operations);
 		return $this;
 	}
 
@@ -246,10 +334,20 @@ class Capability implements Serializable, JsonSerializable
 	 * Allow any operations.
 	 *
 	 * @since 1.0.0
+	 * @since 1.0.1 Return $this
 	 * @return self
 	 */
 	public function allowAny ()
-	{ $this->_operations = ['any']; }
+	{ $this->_operations = ['any']; return $this; }
+
+	/**
+	 * Disallow any operation.
+	 *
+	 * @since 1.0.1
+	 * @return self
+	 */
+	public function disallowAny ()
+	{ $this->_operations = []; return $this; }
 
 	/**
 	 * Check if allow any operation.
@@ -259,24 +357,6 @@ class Capability implements Serializable, JsonSerializable
 	 */
 	public function isAnyAllowed () : bool
 	{ return \in_array('any', $this->_operations, true) !== false; }
-
-	/**
-	 * Remove the any operation from operations.
-	 *
-	 * @since 1.0.0
-	 * @return self
-	 */
-	private function _flushAny ()
-	{
-		$this->_operations = \array_filter(
-			$this->_operations,
-			function ( $operation ) {
-				return $operation !== 'any';
-			}
-		);
-
-		return $this;
-	}
 
 	/**
 	 * Fix spread operation by validating its data and
@@ -300,12 +380,32 @@ class Capability implements Serializable, JsonSerializable
 			}
 			
 			if ( !is_string($item) )
-			{ throw new InvalidArgumentException('Operation term `%s` should be a string or an array.', \gettype($item)); }
+			{ 
+				throw new InvalidArgumentException(
+					\sprintf(
+						'Operation term `%s` should be a string or an array.', 
+						\gettype($item)
+					)
+				); 
+			}
 
 			$_spread[] = $item;
 		}
 
-		return \array_unique($_spread);
+		$_spread = \array_unique($_spread);
+
+		if ( CapabilityOperations::hasInvalid($_spread) )
+		{ 
+			throw new InvalidArgumentException(
+				\sprintf(
+					'Invalid operation term `%s` found. Should be any of: %s.', 
+					\gettype($item), 
+					implode('`, `', CapabilityOperations::operations())
+				)
+			); 
+		}
+
+		return $_spread;
 	}
 
 	/**